requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
nltk
jupyter
feather-format
opencv-contrib-python
youtube-dl
pydub
liac-arff
textstat
vaderSentiment
requests
tensorflow-gpu
keras
pydot
bert-serving-server
bert-serving-client
